requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Cache and Message Queue
redis==5.0.1
celery==5.3.4
kombu==5.3.4
flower==2.0.1

# Object Storage
minio==7.2.0
boto3==1.34.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8
pyotp==2.9.0
qrcode[pil]==7.4.2
ldap3==2.9.1

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
PyMuPDF==1.23.8
python-pptx==0.6.23

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.25.2
scipy==1.11.4

# Audio/Video Processing
ffmpeg-python==0.2.0
librosa==0.10.1
moviepy==1.0.3

# Machine Learning (for watermark algorithms)
scikit-learn==1.3.2
scikit-image==0.22.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
jinja2==3.1.2

# Time Series Database
clickhouse-driver==0.2.6

# QR Code Generation
qrcode[pil]==7.4.2

# PDF Processing
reportlab==4.0.7
pypdf==3.17.1

# Configuration
pyyaml==6.0.1
toml==0.10.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Serialization
orjson==3.9.10
msgpack==1.0.7